// src/app/model/Properties.ts
export interface Properties {
    id: number;
    propertyId: number;
    companyId: number;
    name: string;
    address: string;
    type: string;
    bedrooms: number;
    bathrooms: number;
    size: number;
    purchaseDate: Date
    value: number;
    rent: number;
    createdAt: Date
    updatedAt: Date
}